refactored benchmarks
