+++ tests/Asn1Parser.Benchmark/Program.cs
@@ -1,5 +1,4 @@
 ﻿using BenchmarkDotNet.Configs;
-using BenchmarkDotNet.Reports;
 using BenchmarkDotNet.Running;
 
 namespace Asn1Parser.Benchmark;
@@ -7,7 +6,7 @@ namespace Asn1Parser.Benchmark;
 internal class Program {
     static void Main(String[] args) {
         IConfig config = DefaultConfig.Instance;
-        Summary summary = BenchmarkRunner.Run<Asn1ReaderBenchmark>(config);
-        summary = BenchmarkRunner.Run<Asn1FormatterBenchmark>(config);
+        //Summary summary = BenchmarkRunner.Run<Asn1ReaderBenchmark>(config);
+        var summary = BenchmarkRunner.Run<Asn1FormatterToHexRawBenchmark>(config);
     }
 }
